1 mensagem de commit

+++ src/domain/Price.java
@@ -7,18 +7,18 @@ import java.util.UUID;
 public class Price implements EntityInterface{
 
     private UUID uuid;
-    private BigDecimal price;
+    private Float price;
     private Date date;
 
     public Price() {
     }
 
-    public Price(Date date, BigDecimal price) {
+    public Price(Date date, Float price) {
         this.date = date;
         this.price = price;
     }
 
-    public Price(Date date, BigDecimal price, UUID uuid) {
+    public Price(Date date, Float price, UUID uuid) {
         this.date = date;
         this.price = price;
         this.uuid = uuid;
